refactor SelectEmbedEntityEmptyState
